refactor(commands): split builtin-commands.ts into modular components (Issue #696)
Split the 945-line builtin-commands.ts into separate files by category:
- session-commands.ts: ResetCommand, StatusCommand, HelpCommand (~76 lines)
- node-commands.ts: ListNodesCommand, SwitchNodeCommand, RestartCommand (~88 lines)
- group-commands.ts: Group management commands (~258 lines)
- passive-command.ts: PassiveCommand (~64 lines)
- debug-commands.ts: SetDebugCommand, ShowDebugCommand, ClearDebugCommand (~88 lines)
- schedule-command.ts: ScheduleCommand (~221 lines)
- task-command.ts: TaskCommand (~194 lines)

The builtin-commands.ts now only contains:
- Import statements from modular files
- Re-exports for backward compatibility
- registerDefaultCommands function

This improves:
- Code organization and maintainability
- Single responsibility principle
- Easier navigation and testing

All 41 existing tests pass.
