+++ src/nodes/commands/commands/index.ts
@@ -1,0 +1,35 @@
+/**
+ * Command implementations - Modular command files.
+ *
+ * This module re-exports all command classes for convenient imports.
+ *
+ * Issue #696: 拆分 builtin-commands.ts
+ */
+
+// Session commands
+export { ResetCommand, StatusCommand, HelpCommand } from './session-commands.js';
+
+// Node commands
+export { ListNodesCommand, SwitchNodeCommand, RestartCommand } from './node-commands.js';
+
+// Group commands
+export {
+  CreateGroupCommand,
+  AddGroupMemberCommand,
+  RemoveGroupMemberCommand,
+  ListGroupMembersCommand,
+  ListGroupCommand,
+  DissolveGroupCommand,
+} from './group-commands.js';
+
+// Passive command
+export { PassiveCommand } from './passive-command.js';
+
+// Debug commands
+export { SetDebugCommand, ShowDebugCommand, ClearDebugCommand } from './debug-commands.js';
+
+// Schedule command
+export { ScheduleCommand } from './schedule-command.js';
+
+// Task command
+export { TaskCommand } from './task-command.js';
